Create README.md
# Holiday_Trek

This is a Holiday Trek Interactive Story created by Brianna De La Riva for a 2018 high school Computer Science final. Re-enterring the coding world and exploring Github. This python code has been unchanged since 2018. 

There are several outcomes to this story! Keep playing with different decisions to find them all out. You are trying to make it to your Christmas Gathering Party in time for Christmas. Do you make it to your destination? Or does it end you with wild turns of situations to deal with.

Enjoy!